+++ react_setup/AwesomeProject/components/forms/AdForm.js
@@ -5,7 +5,8 @@ import { EditProfileButton, AdDescriptionAdding } from "../Buttons";
 import { AntDesign, FontAwesome, Ionicons } from "@expo/vector-icons";
 import RadioButton from "../RadioButton";
 import { TouchableOpacity, Text, TextInput, View, Modal } from "react-native";
-import { adCreationStyles, errorStyle, pickerStyle } from "../../shared/Styles";
+import { errorStyle } from "../../shared/Styles";
+import { adCreationStyles } from "../../shared/adsStyles";
 import { Divider } from "react-native-elements";
 import RichTextEditor from "../RichTextEditor";
 import * as ImagePicker from "expo-image-picker";
