+++ react_setup/AwesomeProject/components/forms/LoginForm.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Formik } from "formik";
 import { TouchableOpacity, TextInput, View, Text } from "react-native";
 import * as yup from "yup";
-import { loginStyles } from "../../shared/Styles";
+import { loginStyles } from "../../shared/loginStyles";
 import { LogInButton } from "../Buttons";
 import { useDispatch } from "react-redux";
 import { login } from "../../store/actions/authentication/authenticationActions";
