+++ react_setup/AwesomeProject/screens/LogIn.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { ImageBackground, ActivityIndicator, View, Text } from "react-native";
 import * as Font from "expo-font";
-import { loginStyles } from "../shared/Styles";
+import { loginStyles } from "../shared/loginStyles";
 import LoginForm from "../components/forms/LoginForm";
 import { login } from "../services/AuthService";
 import LocalStorage from "../localStorage";
