+++ react_setup/AwesomeProject/screens/VipAdModal.js
@@ -8,7 +8,8 @@ import {
   Image,
   ActivityIndicator,
 } from "react-native";
-import { adsStyles, vipModalStyles } from "../shared/Styles";
+import { vipModalStyles } from "../shared/Styles";
+import { adsStyles } from "../shared/adsStyles";
 import { Ionicons } from "@expo/vector-icons";
 import { Fontisto, SimpleLineIcons, Feather } from "@expo/vector-icons";
 import { heightPercentageToDP as hp } from "react-native-responsive-screen";
@@ -17,7 +18,7 @@ import { Dimensions } from "react-native";
 import * as Font from "expo-font";
 import HTMLView from "react-native-htmlview";
 import { connect } from "react-redux";
-import { LinearGradient } from 'expo-linear-gradient';
+import { LinearGradient } from "expo-linear-gradient";
 
 const windowWidth = Dimensions.get("window").width;
 const windowHeight = Dimensions.get("window").height;
@@ -57,137 +58,137 @@ export class AdModal extends React.Component {
                   visible={this.state.showModal}
                   style={vipModalStyles.modal}
                 >
-                    <LinearGradient
-                    colors={['#d2bd84', '#fdfcfb', '#d2bd84', '#fdfcfb']}
+                  <LinearGradient
+                    colors={["#d2bd84", "#fdfcfb", "#d2bd84", "#fdfcfb"]}
                     style={vipModalStyles.modalInnerWrap}
                     start={{ x: 0.15, y: 0.2 }}
-                    end={{ x: 0.99, y: 1}}
+                    end={{ x: 0.99, y: 1 }}
                     locations={[0.2, 0.5, 0.72, 1]}
-                    >
-                  <View>
-                    <View style={vipModalStyles.closeButtonContainer}>
-                      <View></View>
-                      <TouchableOpacity
-                        onPress={() => this.props.toggleModal()}
-                      >
-                        <Ionicons
-                          name="md-close"
-                          size={26}
-                          style={vipModalStyles.closeButton}
-                          color="#ededed"
-                        />
-                      </TouchableOpacity>
-                    </View>
-                    <View style={vipModalStyles.centeredWrap}>
-                      <View style={vipModalStyles.titleContainer}>
-                        <View>
-                          <Text style={vipModalStyles.title}>
-                            {this.props.ad.title}
-                          </Text>
-                        </View>
-                        <LinearGradient
-                        colors={['#b8986a', '#362a19']}
-                        style={vipModalStyles.priceContainer}
-                        locations={[0.01, 1]}>
-                        <View>
-                          <Text style={vipModalStyles.price}>
-                            {this.props.ad.price}
-                          </Text>
-                        </View>
-                        </LinearGradient>
-                      </View>
-                      <View style={vipModalStyles.basicUserInfo}>
-                        <View
-                          style={
-                            windowHeight * 0.37 < windowWidth * 0.7
-                              ? vipModalStyles.profileImageBorderHeight
-                              : vipModalStyles.profileImageBorderWidth
-                          }
-                        >
-                          {this.props.ad.image ? (
-                            <Image
-                              style={
-                                windowHeight * 0.37 < windowWidth * 0.7
-                                  ? vipModalStyles.profileImageHeight
-                                  : vipModalStyles.profileImageWidth
-                              }
-                              source={{ uri: this.props.ad.image }}
-                            />
-                          ) : (
-                            <Image
-                              style={
-                                windowHeight * 0.37 < windowWidth * 0.7
-                                  ? vipModalStyles.profileImageHeight
-                                  : vipModalStyles.profileImageWidth
-                              }
-                              source={avatar}
-                            />
-                          )}
-                        </View>
-                        <Text style={vipModalStyles.profileName}>
-                          {this.props.ad.user.firstName +
-                            " " +
-                            this.props.ad.user.lastName}
-                        </Text>
-                        <View style={vipModalStyles.userLocation}>
-                          <SimpleLineIcons
-                            name="location-pin"
-                            size={hp("2.5%")}
-                            color="#000"
-                          />
-                          <Text style={vipModalStyles.location}>
-                            {this.props.ad.user.location.value}
-                          </Text>
-                        </View>
-                        <View style={vipModalStyles.userMail}>
-                          <Fontisto
-                            name="email"
-                            size={hp("2.5%")}
-                            color="#000"
-                          />
-                          <Text style={vipModalStyles.location}>
-                            {this.props.ad.user.email}
-                          </Text>
-                        </View>
-                        <View style={vipModalStyles.userOntact}>
-                          <Feather
-                            name="phone"
-                            size={hp("2.5%")}
-                            color="#000"
-                          />
-                          <Text style={vipModalStyles.location}>
-                            {this.props.ad.user.phoneNumber}
-                          </Text>
-                        </View>
+                  >
+                    <View>
+                      <View style={vipModalStyles.closeButtonContainer}>
+                        <View></View>
                         <TouchableOpacity
-                          onPress={() => {
+                          onPress={() => this.props.toggleModal()}
+                        >
+                          <Ionicons
+                            name="md-close"
+                            size={26}
+                            style={vipModalStyles.closeButton}
+                            color="#ededed"
+                          />
+                        </TouchableOpacity>
+                      </View>
+                      <View style={vipModalStyles.centeredWrap}>
+                        <View style={vipModalStyles.titleContainer}>
+                          <View>
+                            <Text style={vipModalStyles.title}>
+                              {this.props.ad.title}
+                            </Text>
+                          </View>
+                          <LinearGradient
+                            colors={["#b8986a", "#362a19"]}
+                            style={vipModalStyles.priceContainer}
+                            locations={[0.01, 1]}
+                          >
+                            <View>
+                              <Text style={vipModalStyles.price}>
+                                {this.props.ad.price}
+                              </Text>
+                            </View>
+                          </LinearGradient>
+                        </View>
+                        <View style={vipModalStyles.basicUserInfo}>
+                          <View
+                            style={
+                              windowHeight * 0.37 < windowWidth * 0.7
+                                ? vipModalStyles.profileImageBorderHeight
+                                : vipModalStyles.profileImageBorderWidth
+                            }
+                          >
+                            {this.props.ad.image ? (
+                              <Image
+                                style={
+                                  windowHeight * 0.37 < windowWidth * 0.7
+                                    ? vipModalStyles.profileImageHeight
+                                    : vipModalStyles.profileImageWidth
+                                }
+                                source={{ uri: this.props.ad.image }}
+                              />
+                            ) : (
+                              <Image
+                                style={
+                                  windowHeight * 0.37 < windowWidth * 0.7
+                                    ? vipModalStyles.profileImageHeight
+                                    : vipModalStyles.profileImageWidth
+                                }
+                                source={avatar}
+                              />
+                            )}
+                          </View>
+                          <Text style={vipModalStyles.profileName}>
+                            {this.props.ad.user.firstName +
+                              " " +
+                              this.props.ad.user.lastName}
+                          </Text>
+                          <View style={vipModalStyles.userLocation}>
+                            <SimpleLineIcons
+                              name="location-pin"
+                              size={hp("2.5%")}
+                              color="#000"
+                            />
+                            <Text style={vipModalStyles.location}>
+                              {this.props.ad.user.location.value}
+                            </Text>
+                          </View>
+                          <View style={vipModalStyles.userMail}>
+                            <Fontisto
+                              name="email"
+                              size={hp("2.5%")}
+                              color="#000"
+                            />
+                            <Text style={vipModalStyles.location}>
+                              {this.props.ad.user.email}
+                            </Text>
+                          </View>
+                          <View style={vipModalStyles.userOntact}>
+                            <Feather
+                              name="phone"
+                              size={hp("2.5%")}
+                              color="#000"
+                            />
+                            <Text style={vipModalStyles.location}>
+                              {this.props.ad.user.phoneNumber}
+                            </Text>
+                          </View>
+                          <TouchableOpacity
+                            onPress={() => {
                               this.props.navigation.navigate("Profile", {
                                 username: this.props.ad.user.username,
                                 toggleModal: this.props.toggleModal,
                               });
-                            }
-                          }
-                        >
-                          <View style={vipModalStyles.editButton}>
-                            <Text style={vipModalStyles.editButtonText}>
+                            }}
+                          >
+                            <View style={vipModalStyles.editButton}>
+                              <Text style={vipModalStyles.editButtonText}>
                                 Pogledaj profil
-                            </Text>
-                          </View>
-                        </TouchableOpacity>
-                      </View>
-                      <View style={vipModalStyles.description}>
-                        <HTMLView
-                          value={this.props.ad.description}
-                          stylesheet={{
-                            div: {
-                              color: "#000000",
-                              paddingHorizontal: 5,
-                            },
-                          }}
-                        />
+                              </Text>
+                            </View>
+                          </TouchableOpacity>
+                        </View>
+                        <View style={vipModalStyles.description}>
+                          <HTMLView
+                            value={this.props.ad.description}
+                            stylesheet={{
+                              div: {
+                                color: "#000000",
+                                paddingHorizontal: 5,
+                              },
+                            }}
+                          />
+                        </View>
                       </View>
                     </View>
-                  </View>
                   </LinearGradient>
                 </Modal>
               </View>
