+++ react_setup/AwesomeProject/shared/adsStyles.js
@@ -1,0 +1,915 @@
+import { Platform, StyleSheet } from "react-native";
+import { withTheme } from "react-native-elements";
+import {
+  widthPercentageToDP as wp,
+  heightPercentageToDP as hp,
+} from "react-native-responsive-screen";
+
+const customFonts = {
+  "Comfortaa-Regular": require("../assets/fonts/Comfortaa-Regular.ttf"),
+  "Comfortaa-Light": require("../assets/fonts/Comfortaa-Light.ttf"),
+  "Comfortaa-Bold": require("../assets/fonts/Comfortaa-Bold.ttf"),
+  "Comfortaa-Bold": require("./../assets/fonts/Comfortaa-Bold.ttf"),
+  "Roboto-Thin": require("../assets/fonts/Roboto-Thin.ttf"),
+  "Roboto-Bold": require("../assets/fonts/Roboto-Bold.ttf"),
+  "Roboto-Regular": require("../assets/fonts/Roboto-Regular.ttf"),
+  "Roboto-Light": require("../assets/fonts/Roboto-Light.ttf"),
+};
+
+export const adsStyles = StyleSheet.create({
+  backgroundImageContainer: {
+    flex: 1,
+    resizeMode: "cover",
+  },
+  mainContainer: {
+    flex: 1,
+    alignSelf: "center",
+  },
+  titleContainer: {
+    alignContent: "center",
+    alignSelf: "center",
+    marginTop: hp("2.5%"),
+    marginBottom: hp("1%"),
+    width: wp("75%"),
+    height: hp("6.2%"),
+    paddingTop: hp("0.7%"),
+    borderRadius: 20,
+    backgroundColor: "#1e1c24",
+    borderColor: "#ededed",
+  },
+  numberOfAdsContainer: {
+    height: 20,
+  },
+  titleText: {
+    color: "#ffffff",
+    textAlign: "center",
+    paddingTop: Platform.OS === "ios" ? hp("1.6%") : hp("0.8%"),
+    fontSize: wp("6.3%"),
+    lineHeight: wp("7%"),
+    fontFamily: "Comfortaa-Bold",
+  },
+  adContainer: {
+    alignSelf: "center",
+    marginTop: hp("1%"),
+    height: wp("29.92%"),
+    width: wp("90%"),
+  },
+  smallAdContainer: {
+    alignSelf: "center",
+    marginTop: hp("0.5%"),
+    width: wp("90%"),
+  },
+});
+
+
+
+export const adStyles = StyleSheet.create({
+  mainContainer: {
+    backgroundColor: "#2d2d2d",
+    width: "100%",
+    height: "100%",
+    borderRadius: 7,
+    borderWidth: 1,
+    borderColor: "#565656",
+    shadowColor: "#000",
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+
+    elevation: 5,
+  },
+  adMainContainer: {
+    flexDirection: "row",
+  },
+  adImage: {
+    height: wp("27.2%"),
+    width: wp("27.2%"),
+    top: wp("0.68%"),
+    left: wp("0.68%"),
+    borderWidth: 2,
+    borderColor: "#ededed",
+  },
+  adMainText: {
+    marginLeft: wp("2.72%"),
+    width: wp("36.6%"),
+    backgroundColor: "#2d2d2d",
+  },
+  adTitleContainer: {
+    borderBottomWidth: 0.5,
+    borderColor: "#9f9f9f",
+    marginBottom: hp("0.5%"),
+  },
+  adTitle: {
+    paddingTop: Platform.OS === "ios" ? hp("0.5%") : hp("0.25%"),
+    textAlign: "center",
+    fontSize: wp("3%"),
+    fontFamily: "Comfortaa-Bold",
+    color: "#ededed",
+  },
+  descriptionSmall: {
+    width: "100%",
+    height: wp("19%"),
+    borderRadius: 5,
+    borderWidth: 1,
+    borderColor: "#9f9f9f",
+    fontFamily: "Roboto-Regular",
+    backgroundColor: "#1e1c24",
+  },
+  descriptionLarge: {
+    width: "100%",
+    height: wp("15.5%"),
+    borderRadius: 5,
+    borderWidth: 1,
+    borderColor: "#9f9f9f",
+    fontFamily: "Roboto-Regular",
+    backgroundColor: "#1e1c24",
+  },
+  descriptionText: {
+    fontSize: wp("2.5%"),
+    marginLeft: wp("0.85%"),
+    marginRight: wp("0.85%"),
+    color: "#ededed",
+    fontFamily: "Roboto-Regular",
+  },
+  descriptionDetails: {
+    alignSelf: "center",
+  },
+  descriptionDetailsText: {
+    fontSize: wp("2.25%"),
+    fontWeight: "bold",
+    fontFamily: "Roboto-Regular",
+    textDecorationLine: "underline",
+    color: "#ededed",
+  },
+  detailsArrow: {
+    alignSelf: "center",
+    fontSize: wp("2.75%"),
+    color: "#ededed",
+  },
+  ownerNameContainer: {
+    alignSelf: "flex-start",
+    marginTop: wp("0.5%"),
+  },
+  ownerName: {
+    fontSize: wp("2%"),
+    fontFamily: "Roboto-Regular",
+    fontStyle: "italic",
+    color: "#ededed",
+  },
+  adDetails: {
+    marginTop: wp("1%"),
+    marginLeft: wp("1.5%"),
+    width: wp("20.4%"),
+    backgroundColor: "#2d2d2d",
+  },
+  priceContainer: {
+    alignSelf: "center",
+    marginTop: wp("0.3%"),
+    height: wp("5.5%"),
+    width: wp("17.7%"),
+    borderRadius: 5,
+    borderColor: "#7d7d7d",
+    backgroundColor: "#03d5ff",
+  },
+  priceValue: {
+    textAlign: "center",
+    marginTop: wp("1%"),
+    fontSize: wp("2.75%"),
+    fontFamily: "Roboto-Regular",
+    color: "#ededed",
+  },
+  publishDateContainer: {
+    alignContent: "center",
+    alignItems: "center",
+    alignSelf: "center",
+    marginTop: wp("1%"),
+    width: wp("16.4%"),
+    backgroundColor: "#2d2d2d",
+  },
+  publishTitle: {
+    fontSize: wp("2.5%"),
+    fontFamily: "Roboto-Light",
+    color: "#ededed",
+  },
+  publishDate: {
+    fontSize: wp("2.5%"),
+    color: "#ededed",
+  },
+  location: {
+    alignContent: "center",
+    alignItems: "center",
+    alignSelf: "center",
+    width: wp("16.4%"),
+    marginTop: wp("2.75%"),
+    backgroundColor: "#2d2d2d",
+  },
+  locationText: {
+    fontSize: wp("2.5%"),
+    fontFamily: "Roboto-Bold",
+    color: "#ededed",
+  },
+  ratingStars: {
+    flexDirection: "row",
+    justifyContent: "center",
+    width: wp("20.4%"),
+    marginTop: wp("3.5%"),
+  },
+  likeText: {
+    fontSize: wp("2.5%"),
+    fontFamily: "Roboto-Light",
+    color: "#ededed",
+  },
+  like: {
+    marginLeft: wp("1.3%"),
+    fontSize: wp("3%"),
+    color: "#ededed",
+  },
+  dislike: {
+    marginLeft: wp("1.25%"),
+    fontSize: wp("3%"),
+    color: "#ededed",
+  },
+  dislikeText: {
+    marginLeft: wp("1.3%"),
+    fontSize: wp("2.75%"),
+    fontFamily: "Roboto-Light",
+    color: "#ededed",
+  },
+});
+
+export const smallAdStyles = StyleSheet.create({
+  mainContainer: {
+    backgroundColor: "#1e1c24",
+    height: wp("18.72%"),
+    marginHorizontal: wp("1.5%"),
+    marginVertical: hp("0.15%"),
+    borderRadius: 5,
+    borderWidth: wp("0.1%"),
+    borderColor: "#9f9f9f",
+  },
+  adMainContainer: {
+    flexDirection: "row",
+  },
+  adImage: {
+    height: wp("16%"),
+    width: wp("16%"),
+    top: wp("1.36%"),
+    left: wp("1.36%"),
+    borderWidth: 1,
+    borderColor: "#ffffff",
+  },
+  adMainText: {
+    marginTop: wp("1%"),
+    marginLeft: wp("2.72%"),
+    width: wp("50%"),
+  },
+  adTitle: {
+    textAlign: "left",
+    fontSize: wp("3%"),
+    lineHeight: wp("4.4%"),
+    fontFamily: "Comfortaa-Bold",
+    color: "#ededed",
+  },
+  descriptionSmall: {
+    width: "100%",
+    borderTopWidth: 1,
+    borderTopColor: "#2d2d2d",
+    height: wp("12%"),
+  },
+  descriptionText: {
+    fontSize: wp("2.25%"),
+    marginTop: wp("1%"),
+    marginLeft: wp("0.85%"),
+    marginRight: wp("0.85%"),
+    color: "#ededed",
+  },
+  leftDashContainer: {
+    marginLeft: wp("0.3%"),
+    flexDirection: "row",
+  },
+  line: {
+    textAlignVertical: "center",
+    fontSize: 5,
+    color: "#ededed",
+  },
+  adDetails: {
+    marginTop: wp("1.36%"),
+    marginLeft: wp("1%"),
+    width: wp("16%"),
+  },
+  priceContainer: {
+    alignSelf: "center",
+    marginTop: wp("0.3%"),
+    marginBottom: wp("5%"),
+    height: wp("4%"),
+    width: wp("15%"),
+    borderRadius: 5,
+    backgroundColor: "#03d5ff",
+  },
+  priceValue: {
+    textAlign: "center",
+    marginTop: wp("1%"),
+    fontSize: wp("1.75%"),
+    color: "#ededed",
+  },
+  publishDateContainer: {
+    alignContent: "center",
+    alignItems: "center",
+    alignSelf: "center",
+    marginTop: wp("1%"),
+    width: wp("16.4%"),
+    paddingLeft: wp("4%"),
+  },
+  publishTitle: {
+    fontSize: wp("2%"),
+    color: "#ededed",
+  },
+  publishDate: {
+    fontSize: wp("2%"),
+    color: "#ededed",
+  },
+  locationText: {
+    fontSize: wp("2.25%"),
+    color: "#ededed",
+  },
+  ratingStars: {
+    flexDirection: "row",
+    justifyContent: "center",
+    width: wp("15%"),
+    marginTop: wp("3%"),
+  },
+  likeText: {
+    marginLeft: wp("1%"),
+    fontSize: wp("2%"),
+    color: "#ededed",
+  },
+  like: {
+    marginLeft: wp("1.36%"),
+    fontSize: wp("2%"),
+    color: "#ededed",
+  },
+  dislike: {
+    marginLeft: wp("1.36%"),
+    fontSize: wp("2%"),
+    color: "#ededed",
+  },
+  dislikeText: {
+    marginLeft: wp("1.36%"),
+    fontSize: wp("2%"),
+    color: "#ededed",
+  },
+});
+
+export const adCreationStyles = StyleSheet.create({
+  backgroundImageContainer: {
+    flex: 1,
+    resizeMode: "cover",
+  },
+  mainContainer: {
+    alignSelf: "center",
+    marginTop: hp("2%"),
+    width: wp("90%"),
+    minHeight: hp("80%"),
+    borderRadius: 10,
+    backgroundColor: "#2d2d2d",
+  },
+  inputFieldContainer: {
+    alignSelf: "center",
+    marginLeft: wp("1%"),
+    width: wp("82%"),
+  },
+  fieldName: {
+    marginLeft: wp("1%"),
+    marginTop: hp("1.25%"),
+    fontSize: wp("3.15%"),
+    fontFamily: "Roboto-Light",
+    color: "#ededed",
+  },
+  adNameField: {
+    marginTop: hp("0.75%"),
+    width: wp("82%"),
+    height: hp("6%"),
+    paddingLeft: wp("2.5%"),
+    fontSize: wp("3.5%"),
+    fontFamily: "Roboto-Light",
+    fontStyle: "italic",
+    color: "white",
+    opacity: 0.8,
+    backgroundColor: "#1e1c24",
+    borderRadius: 8,
+    borderWidth: 1,
+    borderColor: "white",
+  },
+  adDescriptionField: {
+    width: wp("82%"),
+    marginTop: hp("0.75%"),
+    paddingLeft: wp("1.5%"),
+    fontSize: wp("3.45%"),
+    fontFamily: "Roboto-Light",
+    fontStyle: "italic",
+    color: "#ededed",
+    opacity: 0.8,
+    backgroundColor: "#1e1c24",
+    borderRadius: 10,
+    borderWidth: 1,
+    borderColor: "white",
+    height: hp("5%"),
+  },
+  dropDownCatSubContainer: {
+    marginTop: hp("0.75%"),
+    borderRadius: 10,
+    borderWidth: 1,
+    borderColor: "white",
+    backgroundColor: "#1e1c24",
+    height: hp("16%"),
+  },
+  dropDownCatContainer: {
+    alignSelf: "center",
+    marginTop: hp("0.25%"),
+    width: wp("75%"),
+    backgroundColor: "#1e1c24",
+    height: hp("7.5%"),
+  },
+  dropDownSubContainer: {
+    alignSelf: "center",
+    marginTop: hp("0.25%"),
+    width: wp("75%"),
+    backgroundColor: "#1e1c24",
+    height: hp("7.5%"),
+    fontWeight: "bold",
+  },
+  priceRBContainer: {
+    marginTop: hp("0.75%"),
+    borderRadius: 10,
+    borderWidth: 1,
+    borderColor: "white",
+    backgroundColor: "#1e1c24",
+    height: hp("9%"),
+  },
+  pickImageContainer: {
+    alignSelf: "center",
+    marginTop: hp("0.75%"),
+    width: wp("82%"),
+    borderRadius: 8,
+    borderWidth: 1,
+    borderColor: "white",
+    backgroundColor: "#1e1c24",
+    opacity: 0.8,
+    height: hp("6%"),
+  },
+  pickImageAdditional: {
+    alignSelf: "center",
+    flexDirection: "row",
+    justifyContent: "space-between",
+    width: wp("75%"),
+  },
+  pickingImage: {
+    paddingTop: hp("1.25%"),
+    fontSize: wp("4%"),
+    fontFamily: "Roboto-Light",
+    color: "#ededed",
+    fontStyle: "italic",
+  },
+  plusIcon: {
+    fontSize: wp("4%"),
+    color: "#ededed",
+    paddingTop: hp("1.75%"),
+  },
+  inputFieldAdditionalContainer: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginLeft: wp("1%"),
+    width: wp("82%"),
+  },
+  questionMarkIcon: {
+    marginLeft: wp("1%"),
+    fontSize: wp("3.75%"),
+    color: "#ededed",
+    paddingTop: hp("1.5%"),
+  },
+  centeredView: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+    marginTop: 22,
+  },
+  modalView: {
+    margin: 20,
+    backgroundColor: "white",
+    borderRadius: 20,
+    padding: 35,
+    alignItems: "center",
+    shadowColor: "#000",
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 4,
+    elevation: 5,
+    height: hp("82%"),
+    width: wp("78%"),
+  },
+  closeIcon: {
+    position: "absolute",
+    right: 5,
+    top: 5,
+  },
+  adTypeTextName: {
+    textAlign: "center",
+    marginLeft: wp("1%"),
+    fontSize: wp("4.25%"),
+    fontWeight: "bold",
+  },
+  adTypeText: {
+    textAlign: "justify",
+    marginLeft: wp("0.5%"),
+    fontSize: wp("3.75%"),
+    paddingTop: hp("1%"),
+    paddingBottom: hp("1%"),
+  },
+  adNoteText: {
+    textAlign: "justify",
+    marginLeft: wp("0.5%"),
+    fontSize: wp("3.75%"),
+    paddingTop: hp("1.5%"),
+    fontWeight: "bold",
+  },
+  dropDownTypeWrapper: {
+    flexDirection: "row",
+    alignSelf: "center",
+    justifyContent: "space-between",
+    width: wp("82%"),
+  },
+  dropDownTypeContainer: {
+    marginTop: hp("0.75%"),
+    width: wp("45%"),
+    borderRadius: 8,
+    borderWidth: 1,
+    borderColor: "white",
+    backgroundColor: "#1e1c24",
+    height: hp("8%"),
+  },
+  typeCodeInput: {
+    alignSelf: "center",
+    textAlign: "center",
+    marginTop: hp("0.75%"),
+    width: wp("35%"),
+    height: hp("8%"),
+    fontSize: wp("3.5%"),
+    fontFamily: "Roboto-Light",
+    fontStyle: "italic",
+    color: "white",
+    opacity: 0.8,
+    backgroundColor: "#1e1c24",
+    borderRadius: 8,
+    borderWidth: 1,
+    borderColor: "white",
+    opacity: 0.8,
+  },
+});
+
+export const vipAdStyles = StyleSheet.create({
+  mainContainer: {
+    width: "100%",
+    height: "100%",
+    borderRadius: 7,
+    borderWidth: 1,
+    borderColor: "#565656",
+    shadowColor: "#000",
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+
+    elevation: 5,
+  },
+  adMainContainer: {
+    flexDirection: "row",
+  },
+  adImage: {
+    height: wp("27.2%"),
+    width: wp("27.2%"),
+    top: wp("0.68%"),
+    left: wp("0.68%"),
+    borderWidth: 2,
+    borderColor: "#ededed",
+  },
+  adMainText: {
+    marginLeft: wp("2.72%"),
+    width: wp("36.6%"),
+  },
+  adTitleContainer: {
+    borderBottomWidth: 0.5,
+    borderColor: "#9f9f9f",
+    marginBottom: hp("0.5%"),
+  },
+  adTitle: {
+    paddingTop: Platform.OS === "ios" ? hp("0.5%") : hp("0.25%"),
+    textAlign: "center",
+    fontSize: wp("3%"),
+    fontFamily: "Comfortaa-Bold",
+    color: "black",
+  },
+  descriptionSmall: {
+    width: "100%",
+    height: wp("19%"),
+    borderRadius: 5,
+    borderWidth: 1,
+    borderColor: "#9f9f9f",
+    fontFamily: "Roboto-Regular",
+    backgroundColor: "white",
+  },
+  descriptionLarge: {
+    width: "100%",
+    height: wp("15.5%"),
+    borderRadius: 5,
+    borderWidth: 1,
+    borderColor: "#9f9f9f",
+    fontFamily: "Roboto-Regular",
+    backgroundColor: "white",
+  },
+  descriptionText: {
+    fontSize: wp("2.5%"),
+    marginLeft: wp("0.85%"),
+    marginRight: wp("0.85%"),
+    color: "black",
+    fontFamily: "Roboto-Regular",
+  },
+  descriptionDetails: {
+    alignSelf: "center",
+  },
+  descriptionDetailsText: {
+    fontSize: wp("2.25%"),
+    fontWeight: "bold",
+    fontFamily: "Roboto-Regular",
+    textDecorationLine: "underline",
+    color: "black",
+  },
+  detailsArrow: {
+    alignSelf: "center",
+    fontSize: wp("2.75%"),
+    color: "black",
+  },
+  ownerNameContainer: {
+    alignSelf: "flex-start",
+    marginTop: wp("0.5%"),
+  },
+  ownerName: {
+    fontSize: wp("2%"),
+    fontFamily: "Roboto-Regular",
+    fontStyle: "italic",
+    color: "black",
+  },
+  adDetails: {
+    marginTop: wp("1%"),
+    marginLeft: wp("1.5%"),
+    width: wp("20.4%"),
+  },
+  priceContainer: {
+    alignSelf: "center",
+    marginTop: wp("0.3%"),
+    height: wp("5.5%"),
+    width: wp("17.7%"),
+    borderRadius: 5,
+    borderColor: "#7d7d7d",
+    backgroundColor: "#03d5ff",
+  },
+  priceValue: {
+    textAlign: "center",
+    marginTop: wp("1%"),
+    fontSize: wp("2.75%"),
+    fontFamily: "Roboto-Regular",
+    color: "white",
+  },
+  publishDateContainer: {
+    alignContent: "center",
+    alignItems: "center",
+    alignSelf: "center",
+    marginTop: wp("5%"),
+    width: wp("16.4%"),
+  },
+  publishTitle: {
+    fontSize: wp("2.5%"),
+    fontFamily: "Roboto-Light",
+    color: "black",
+  },
+  publishDate: {
+    fontSize: wp("2.5%"),
+    color: "black",
+  },
+  location: {
+    alignContent: "center",
+    alignItems: "center",
+    alignSelf: "center",
+    width: wp("16.4%"),
+    marginTop: wp("5%"),
+  },
+  locationText: {
+    fontSize: wp("2.5%"),
+    fontFamily: "Roboto-Bold",
+    color: "black",
+  },
+  ratingStars: {
+    flexDirection: "row",
+    justifyContent: "center",
+    width: wp("20.4%"),
+    marginTop: wp("3%"),
+  },
+  likeText: {
+    fontSize: wp("2.5%"),
+    fontFamily: "Roboto-Light",
+    color: "black",
+  },
+  like: {
+    marginLeft: wp("1.3%"),
+    fontSize: wp("3%"),
+    color: "black",
+  },
+  dislike: {
+    marginLeft: wp("1.25%"),
+    fontSize: wp("3%"),
+    color: "black",
+  },
+  dislikeText: {
+    marginLeft: wp("1.3%"),
+    fontSize: wp("2.75%"),
+    fontFamily: "Roboto-Light",
+    color: "black",
+  },
+});
+export const premiumAdStyles = StyleSheet.create({
+  mainContainer: {
+    width: "100%",
+    height: "100%",
+    borderRadius: 7,
+    borderWidth: 1,
+    borderColor: "#565656",
+    shadowColor: "#000",
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+
+    elevation: 5,
+  },
+  adMainContainer: {
+    flexDirection: "row",
+  },
+  adImage: {
+    height: wp("27.2%"),
+    width: wp("27.2%"),
+    top: wp("0.68%"),
+    left: wp("0.68%"),
+    borderWidth: 2,
+    borderColor: "#ededed",
+  },
+  adMainText: {
+    marginLeft: wp("2.72%"),
+    width: wp("36.6%"),
+  },
+  adTitleContainer: {
+    borderBottomWidth: 0.5,
+    borderColor: "#9f9f9f",
+    marginBottom: hp("0.5%"),
+  },
+  adTitle: {
+    paddingTop: Platform.OS === "ios" ? hp("0.5%") : hp("0.25%"),
+    textAlign: "center",
+    fontSize: wp("3%"),
+    fontFamily: "Comfortaa-Bold",
+    color: "black",
+  },
+  descriptionSmall: {
+    width: "100%",
+    height: wp("19%"),
+    borderRadius: 5,
+    borderWidth: 1,
+    borderColor: "#9f9f9f",
+    fontFamily: "Roboto-Regular",
+    backgroundColor: "white",
+  },
+  descriptionLarge: {
+    width: "100%",
+    height: wp("15.5%"),
+    borderRadius: 5,
+    borderWidth: 1,
+    borderColor: "#9f9f9f",
+    fontFamily: "Roboto-Regular",
+    backgroundColor: "white",
+  },
+  descriptionText: {
+    fontSize: wp("2.5%"),
+    marginLeft: wp("0.85%"),
+    marginRight: wp("0.85%"),
+    color: "black",
+    fontFamily: "Roboto-Regular",
+  },
+  descriptionDetails: {
+    alignSelf: "center",
+  },
+  descriptionDetailsText: {
+    fontSize: wp("2.25%"),
+    fontWeight: "bold",
+    fontFamily: "Roboto-Regular",
+    textDecorationLine: "underline",
+    color: "black",
+  },
+  detailsArrow: {
+    alignSelf: "center",
+    fontSize: wp("2.75%"),
+    color: "black",
+  },
+  ownerNameContainer: {
+    alignSelf: "flex-start",
+    marginTop: wp("0.5%"),
+  },
+  ownerName: {
+    fontSize: wp("2%"),
+    fontFamily: "Roboto-Regular",
+    fontStyle: "italic",
+    color: "black",
+  },
+  adDetails: {
+    marginTop: wp("1%"),
+    marginLeft: wp("1.5%"),
+    width: wp("20.4%"),
+  },
+  priceContainer: {
+    alignSelf: "center",
+    marginTop: wp("0.3%"),
+    height: wp("5.5%"),
+    width: wp("17.7%"),
+    borderRadius: 5,
+    borderColor: "#7d7d7d",
+  },
+  priceValue: {
+    textAlign: "center",
+    marginTop: wp("1%"),
+    fontSize: wp("2.75%"),
+    fontFamily: "Roboto-Regular",
+    color: "white",
+  },
+  publishDateContainer: {
+    alignContent: "center",
+    alignItems: "center",
+    alignSelf: "center",
+    marginTop: wp("5%"),
+    width: wp("16.4%"),
+  },
+  publishTitle: {
+    fontSize: wp("2.5%"),
+    fontFamily: "Roboto-Light",
+    color: "black",
+  },
+  publishDate: {
+    fontSize: wp("2.5%"),
+    color: "black",
+  },
+  location: {
+    alignContent: "center",
+    alignItems: "center",
+    alignSelf: "center",
+    width: wp("16.4%"),
+    marginTop: wp("5%"),
+  },
+  locationText: {
+    fontSize: wp("2.5%"),
+    fontFamily: "Roboto-Bold",
+    color: "black",
+  },
+  ratingStars: {
+    flexDirection: "row",
+    justifyContent: "center",
+    width: wp("20.4%"),
+    marginTop: wp("3%"),
+  },
+  likeText: {
+    fontSize: wp("2.5%"),
+    fontFamily: "Roboto-Light",
+    color: "black",
+  },
+  like: {
+    marginLeft: wp("1.3%"),
+    fontSize: wp("3%"),
+    color: "black",
+  },
+  dislike: {
+    marginLeft: wp("1.25%"),
+    fontSize: wp("3%"),
+    color: "black",
+  },
+  dislikeText: {
+    marginLeft: wp("1.3%"),
+    fontSize: wp("2.75%"),
+    fontFamily: "Roboto-Light",
+    color: "black",
+  },
+});
