+++ react_setup/AwesomeProject/shared/contactUsStyles.js
@@ -1,0 +1,183 @@
+import { Platform, StyleSheet } from "react-native";
+import { withTheme } from "react-native-elements";
+import {
+  widthPercentageToDP as wp,
+  heightPercentageToDP as hp,
+} from "react-native-responsive-screen";
+
+const customFonts = {
+  "Comfortaa-Regular": require("../assets/fonts/Comfortaa-Regular.ttf"),
+  "Comfortaa-Light": require("../assets/fonts/Comfortaa-Light.ttf"),
+  "Comfortaa-Bold": require("../assets/fonts/Comfortaa-Bold.ttf"),
+  "Comfortaa-Bold": require("./../assets/fonts/Comfortaa-Bold.ttf"),
+  "Roboto-Thin": require("../assets/fonts/Roboto-Thin.ttf"),
+  "Roboto-Bold": require("../assets/fonts/Roboto-Bold.ttf"),
+  "Roboto-Regular": require("../assets/fonts/Roboto-Regular.ttf"),
+  "Roboto-Light": require("../assets/fonts/Roboto-Light.ttf"),
+};
+
+export const contactUsStyles = StyleSheet.create({
+  backgroundImageContainer: {
+    flex: 1,
+    resizeMode: "cover",
+  },
+  mainContainer: {
+    alignSelf: "center",
+    marginTop: hp("2%"),
+    width: wp("90%"),
+    minHeight: hp("76%"),
+    borderRadius: 10,
+    backgroundColor: "#2d2d2d",
+  },
+  backForwardContainer: {
+    height: hp("2%"),
+  },
+  inputFieldContainer: {
+    marginTop: hp("1%"),
+    alignSelf: "center",
+    marginLeft: wp("1%"),
+    width: wp("82%"),
+  },
+  inputFieldContainerWithMargin: {
+    alignSelf: "center",
+    marginLeft: wp("1%"),
+    marginTop: hp("2%"),
+    width: wp("82%"),
+  },
+  fieldName: {
+    marginLeft: wp("1%"),
+    marginTop: hp("1.25%"),
+    fontSize: wp("3.15%"),
+    fontFamily: "Roboto-Light",
+    color: "#ededed",
+  },
+  fieldNameBold: {
+    marginLeft: wp("1%"),
+    marginTop: hp("1%"),
+    fontSize: wp("3.5%"),
+    fontFamily: "Roboto-Bold",
+    color: "#ededed",
+  },
+  adNameField: {
+    marginTop: hp("0.75%"),
+    width: wp("82%"),
+    height: hp("4.5%"),
+    paddingLeft: wp("1.5%"),
+    fontSize: wp("3.75%"),
+    fontFamily: "Roboto-Light",
+    fontStyle: "italic",
+    color: "white",
+    opacity: 0.8,
+    backgroundColor: "#1e1c24",
+    borderRadius: 8,
+    borderWidth: 1,
+    borderColor: "white",
+  },
+  adDescriptionField: {
+    width: wp("82%"),
+    marginTop: hp("0.75%"),
+    paddingLeft: wp("1.5%"),
+    fontSize: wp("3.75%"),
+    fontFamily: "Roboto-Light",
+    color: "#ededed",
+    fontStyle: "italic",
+    opacity: 0.8,
+    backgroundColor: "#1e1c24",
+    borderRadius: 10,
+    borderWidth: 1,
+    borderColor: "white",
+    maxHeight: hp("16%"),
+  },
+  dropDownCatSubContainer: {
+    marginTop: hp("0.75%"),
+    borderRadius: 10,
+    borderWidth: 1,
+    borderColor: "white",
+    backgroundColor: "#1e1c24",
+    height: hp("9%"),
+  },
+  dropDownCatContainer: {
+    marginTop: hp("0.25%"),
+    alignSelf: "center",
+    width: wp("75%"),
+    borderBottomWidth: 1,
+    borderColor: "white",
+    backgroundColor: "#1e1c24",
+    height: hp("4%"),
+  },
+  dropDownSubContainer: {
+    alignSelf: "center",
+    width: wp("75%"),
+    backgroundColor: "#1e1c24",
+    height: hp("4%"),
+    fontWeight: "bold",
+  },
+  priceRBContainer: {
+    marginTop: hp("0.75%"),
+    borderRadius: 10,
+    borderWidth: 1,
+    borderColor: "white",
+    backgroundColor: "#1e1c24",
+    height: hp("9%"),
+  },
+  pickImageContainer: {
+    alignSelf: "center",
+    marginTop: hp("0.75%"),
+    width: wp("82%"),
+    borderRadius: 8,
+    borderWidth: 1,
+    borderColor: "white",
+    opacity: 0.8,
+    backgroundColor: "#1e1c24",
+    height: hp("4.5%"),
+  },
+  pickImageAdditional: {
+    alignSelf: "center",
+    flexDirection: "row",
+    justifyContent: "space-between",
+    width: wp("75%"),
+  },
+  pickingImage: {
+    paddingTop: hp("0.65%"),
+    fontSize: wp("3.75%"),
+    fontFamily: "Roboto-Light",
+    color: "#ededed",
+    fontStyle: "italic",
+  },
+  plusIcon: {
+    fontSize: wp("3.75%"),
+    color: "#ededed",
+    paddingTop: hp("0.85%"),
+  },
+  inputFieldAdditionalContainer: {
+    flexDirection: "row",
+  },
+  questionMarkIcon: {
+    marginLeft: wp("1%"),
+    fontSize: wp("3.75%"),
+    color: "#ededed",
+    paddingTop: hp("1.25%"),
+  },
+  dropDownTypeContainer: {
+    marginTop: hp("0.75%"),
+    borderRadius: 8,
+    borderWidth: 1,
+    borderColor: "white",
+    backgroundColor: "#1e1c24",
+    height: hp("4.5%"),
+  },
+  typeCodeInput: {
+    alignSelf: "center",
+    textAlign: "center",
+    marginTop: hp("0.75%"),
+    width: wp("40%"),
+    height: hp("4%"),
+    paddingLeft: wp("1.5%"),
+    borderWidth: 1,
+    borderRadius: 8,
+    borderColor: "#ededed",
+    fontSize: wp("3.5%"),
+    fontStyle: "italic",
+    opacity: 0.8,
+  },
+});
