+++ react_setup/AwesomeProject/shared/modalStyles.js
@@ -1,0 +1,782 @@
+import { Platform, StyleSheet } from "react-native";
+import { withTheme } from "react-native-elements";
+import {
+  widthPercentageToDP as wp,
+  heightPercentageToDP as hp,
+} from "react-native-responsive-screen";
+
+const customFonts = {
+  "Comfortaa-Regular": require("../assets/fonts/Comfortaa-Regular.ttf"),
+  "Comfortaa-Light": require("../assets/fonts/Comfortaa-Light.ttf"),
+  "Comfortaa-Bold": require("../assets/fonts/Comfortaa-Bold.ttf"),
+  "Comfortaa-Bold": require("./../assets/fonts/Comfortaa-Bold.ttf"),
+  "Roboto-Thin": require("../assets/fonts/Roboto-Thin.ttf"),
+  "Roboto-Bold": require("../assets/fonts/Roboto-Bold.ttf"),
+  "Roboto-Regular": require("../assets/fonts/Roboto-Regular.ttf"),
+  "Roboto-Light": require("../assets/fonts/Roboto-Light.ttf"),
+};
+
+export const modalStyles = StyleSheet.create({
+  blurView: {
+    width: wp("100%"),
+    height: hp("100%"),
+  },
+  modalWrap: {
+    alignSelf: "center",
+    marginTop: hp("1%"),
+  },
+  modal: {
+    borderWidth: 0,
+    borderColor: "#d6d6d6",
+  },
+  modalInnerWrap: {
+    marginTop: hp("3%"),
+    alignSelf: "center",
+    width: wp("90%"),
+    height: hp("90%"),
+    borderRadius: 5,
+    backgroundColor: "#2d2d2d",
+  },
+  modalInnerWrapComment: {
+    backgroundColor: "#ededed",
+    marginTop: hp("3%"),
+    alignSelf: "center",
+    width: wp("90%"),
+    height: hp("89%"),
+    borderRadius: 5,
+  },
+  closeButtonContainer: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+  },
+  closeButton: {
+    marginTop: hp("0.5%"),
+    marginRight: hp("0.5%"),
+  },
+  titleContainer: {
+    backgroundColor: "#1e1c24",
+    borderRadius: 10,
+    width: wp("86%"),
+    minHeight: hp("5.25%"),
+    marginLeft: wp("2.5%"),
+    marginRight: wp("2.5%"),
+    marginBottom: hp("0.25%"),
+    textAlign: "center",
+    flexDirection: "row",
+    flex: 1,
+    justifyContent: "space-between",
+  },
+  title: {
+    fontSize: hp("2.5%"),
+    color: "#ededed",
+    fontFamily: "Roboto-Bold",
+    marginTop: hp("0.65%"),
+    marginLeft: wp("2%"),
+  },
+  priceContainer: {
+    height: hp("3%"),
+    width: wp("20%"),
+    backgroundColor: "#03d5ff",
+    marginTop: hp("0.9%"),
+    marginRight: wp("1.5%"),
+    borderRadius: 5,
+  },
+  price: {
+    marginTop: hp("0.25%"),
+    color: "#ededed",
+    fontSize: hp("1.75%"),
+    fontFamily: "Roboto-Bold",
+    textAlign: "center",
+  },
+  centerdWrap: {
+    alignSelf: "center",
+  },
+  basicUserInfo: {
+    maxHeight: hp("50%"),
+    marginBottom: wp("1%"),
+    marginTop: hp("1%"),
+    marginLeft: wp("2.5%"),
+    marginRight: wp("2.5%"),
+    backgroundColor: "#1e1c24",
+    borderRadius: 10,
+  },
+  profileImageBorderHeight: {
+    alignSelf: "center",
+    marginTop: hp("1.5%"),
+    width: hp("16%"),
+    height: hp("16%"),
+    borderWidth: 1,
+    borderColor: "#ffffff",
+  },
+  profileImageBorderWidth: {
+    alignSelf: "center",
+    marginTop: wp("3%"),
+    width: wp("32%"),
+    height: wp("32%"),
+    borderWidth: 1,
+    borderColor: "#ffffff",
+  },
+  mainContainer: {
+    alignSelf: "center",
+    marginTop: hp("3%"),
+    width: wp("90%"),
+    height: wp("60%"),
+  },
+  description: {
+    minHeight: hp("37%"),
+    marginBottom: hp("0.5%"),
+    marginTop: hp("1%"),
+    marginLeft: wp("2.5%"),
+    marginRight: wp("2.5%"),
+    paddingLeft: wp("2%"),
+    paddingLeft: wp("1%"),
+    paddingTop: wp("2%"),
+    paddingBottom: wp("2%"),
+    backgroundColor: "#1e1c24",
+    borderRadius: 10,
+  },
+  profileImageHeight: {
+    alignSelf: "center",
+    marginTop: hp("0.35%"),
+    width: hp("15%"),
+    height: hp("15%"),
+  },
+  profileImageWidth: {
+    alignSelf: "center",
+    marginTop: wp("0.7%"),
+    width: wp("30%"),
+    height: wp("30%"),
+  },
+  profileName: {
+    alignSelf: "center",
+    marginTop: hp("0.25%"),
+    marginBottom: hp("0.5%"),
+    fontSize: hp("2.75%"),
+    fontWeight: "bold",
+    fontFamily: "Roboto-Light",
+    color: "#ededed",
+  },
+  userLocation: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginBottom: hp("0.25%"),
+  },
+  location: {
+    marginLeft: wp("1%"),
+    fontSize: wp("4%"),
+    fontFamily: "Roboto-Light",
+    color: "#ededed",
+  },
+  userMail: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginBottom: hp("0.5%"),
+  },
+  userOntact: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginBottom: hp("1.5%"),
+  },
+  userRating: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginBottom: hp("1.5%"),
+  },
+  ratingText: {
+    fontSize: hp("2%"),
+    fontFamily: "Roboto-Light",
+    color: "#ededed",
+    marginLeft: wp("1%"),
+    marginRight: wp("3%"),
+  },
+  likeComponent: {
+    color: "#ededed",
+  },
+  like: {
+    fontSize: hp("2.75%"),
+    color: "#ededed",
+  },
+  dislikeComponent: {
+    color: "#ededed",
+  },
+  dislike: {
+    fontSize: hp("2.75%"),
+    color: "#ededed",
+  },
+  textInput: {
+    padding: wp("5%"),
+    height: hp("33%"),
+    width: wp("85%"),
+    marginLeft: wp("2.5%"),
+    borderRadius: 10,
+    backgroundColor: "white",
+    fontSize: hp("2%"),
+    fontFamily: "Roboto-Light",
+  },
+  nickname: {
+    paddingLeft: wp("5%"),
+    height: hp("7%"),
+    width: wp("85%"),
+    marginLeft: wp("2.5%"),
+    marginBottom: hp("2%"),
+    borderRadius: 10,
+    backgroundColor: "white",
+    fontSize: hp("2%"),
+    fontFamily: "Roboto-Light",
+  },
+  button: {
+    height: hp("7%"),
+    width: wp("85%"),
+    marginLeft: wp("2.5%"),
+    marginBottom: hp("2%"),
+    marginTop: hp("6%"),
+    borderRadius: 10,
+    backgroundColor: "#1e1c24",
+  },
+  buttonText: {
+    marginTop: hp("1%"),
+    textAlign: "center",
+    fontSize: hp("3%"),
+    fontFamily: "Roboto-Bold",
+    color: "#ededed",
+  },
+  commentHeadingContainer: {
+    marginLeft: wp("2.5%"),
+    marginBottom: hp("0.2%"),
+  },
+  commentHeading: {
+    fontSize: hp("2.5%"),
+    fontFamily: "Roboto-Light",
+  },
+  innerText: {
+    fontSize: hp("1.5%"),
+    fontFamily: "Roboto-Light",
+    marginBottom: hp("0.2%"),
+  },
+  ratingContainer: {
+    marginLeft: wp("25%"),
+    marginRight: wp("25%"),
+    marginBottom: hp("3%"),
+    flexDirection: "row",
+    justifyContent: "space-between",
+  },
+  reviewTitleText: {
+    marginBottom: hp("2%"),
+    fontSize: hp("4%"),
+    fontFamily: "Roboto-Bold",
+    textAlign: "center",
+  },
+  editButton: {
+    height: hp("5%"),
+    borderBottomLeftRadius: 10,
+    borderBottomRightRadius: 10,
+    textAlign: "center",
+    backgroundColor: "#03d5ff",
+    borderColor: "#ededed",
+  },
+  editButtonText: {
+    marginTop: hp("1%"),
+    color: "#ededed",
+    fontSize: hp("2%"),
+    fontWeight: "bold",
+    textAlign: "center",
+  },
+});
+
+export const vipModalStyles = StyleSheet.create({
+  blurView: {
+    width: wp("100%"),
+    height: hp("100%"),
+  },
+  modalWrap: {
+    alignSelf: "center",
+    marginTop: hp("1%"),
+  },
+  modal: {
+    borderWidth: 0,
+    borderColor: "#d6d6d6",
+  },
+  modalInnerWrap: {
+    marginTop: hp("3%"),
+    alignSelf: "center",
+    width: wp("90%"),
+    height: hp("90%"),
+    borderRadius: 5,
+  },
+  modalInnerWrapComment: {
+    backgroundColor: "#ededed",
+    marginTop: hp("3%"),
+    alignSelf: "center",
+    width: wp("90%"),
+    height: hp("89%"),
+    borderRadius: 5,
+  },
+  closeButtonContainer: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+  },
+  closeButton: {
+    marginTop: hp("0.5%"),
+    marginRight: hp("0.5%"),
+  },
+  titleContainer: {
+    backgroundColor: "#1e1c24",
+    borderRadius: 10,
+    width: wp("86%"),
+    minHeight: hp("5.25%"),
+    marginLeft: wp("2.5%"),
+    marginRight: wp("2.5%"),
+    marginBottom: hp("0.25%"),
+    textAlign: "center",
+    flexDirection: "row",
+    flex: 1,
+    justifyContent: "space-between",
+  },
+  title: {
+    fontSize: hp("2.5%"),
+    color: "#ededed",
+    fontFamily: "Roboto-Bold",
+    marginTop: hp("0.65%"),
+    marginLeft: wp("2%"),
+  },
+  priceContainer: {
+    height: hp("3%"),
+    width: wp("20%"),
+    marginTop: hp("0.9%"),
+    marginRight: wp("1.5%"),
+    borderRadius: 5,
+  },
+  price: {
+    marginTop: hp("0.25%"),
+    color: "#ededed",
+    fontSize: hp("1.75%"),
+    fontFamily: "Roboto-Bold",
+    textAlign: "center",
+  },
+  centerdWrap: {
+    alignSelf: "center",
+  },
+  basicUserInfo: {
+    maxHeight: hp("50%"),
+    marginBottom: wp("1%"),
+    marginTop: hp("1%"),
+    marginLeft: wp("2.5%"),
+    marginRight: wp("2.5%"),
+    backgroundColor: "#ffffff",
+    borderRadius: 10,
+  },
+  profileImageBorderHeight: {
+    alignSelf: "center",
+    marginTop: hp("1.5%"),
+    width: hp("16%"),
+    height: hp("16%"),
+    borderWidth: 1,
+    borderColor: "#1e1c24",
+  },
+  profileImageBorderWidth: {
+    alignSelf: "center",
+    marginTop: wp("3%"),
+    width: wp("32%"),
+    height: wp("32%"),
+    borderWidth: 1,
+    borderColor: "#1e1c24",
+  },
+  mainContainer: {
+    alignSelf: "center",
+    marginTop: hp("3%"),
+    width: wp("90%"),
+    height: wp("60%"),
+  },
+  description: {
+    minHeight: hp("37%"),
+    marginBottom: hp("0.5%"),
+    marginTop: hp("1%"),
+    marginLeft: wp("2.5%"),
+    marginRight: wp("2.5%"),
+    paddingLeft: wp("2%"),
+    paddingLeft: wp("1%"),
+    paddingTop: wp("2%"),
+    paddingBottom: wp("2%"),
+    backgroundColor: "#ffffff",
+    borderRadius: 10,
+  },
+  profileImageHeight: {
+    alignSelf: "center",
+    marginTop: hp("0.35%"),
+    width: hp("15%"),
+    height: hp("15%"),
+  },
+  profileImageWidth: {
+    alignSelf: "center",
+    marginTop: wp("0.7%"),
+    width: wp("30%"),
+    height: wp("30%"),
+  },
+  profileName: {
+    alignSelf: "center",
+    marginTop: hp("0.25%"),
+    marginBottom: hp("0.5%"),
+    fontSize: hp("2.75%"),
+    fontWeight: "bold",
+    fontFamily: "Roboto-Light",
+    color: "#000000",
+  },
+  userLocation: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginBottom: hp("0.25%"),
+  },
+  location: {
+    marginLeft: wp("1%"),
+    fontSize: wp("4%"),
+    fontFamily: "Roboto-Light",
+    color: "#000000",
+  },
+  userMail: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginBottom: hp("0.5%"),
+  },
+  userOntact: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginBottom: hp("1.5%"),
+  },
+  userRating: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginBottom: hp("1.5%"),
+  },
+  ratingText: {
+    fontSize: hp("2%"),
+    fontFamily: "Roboto-Light",
+    color: "#000000",
+    marginLeft: wp("1%"),
+    marginRight: wp("3%"),
+  },
+  likeComponent: {
+    color: "#000000",
+  },
+  like: {
+    fontSize: hp("2.75%"),
+    color: "#000000",
+  },
+  dislikeComponent: {
+    color: "#000000",
+  },
+  dislike: {
+    fontSize: hp("2.75%"),
+    color: "#000000",
+  },
+  textInput: {
+    padding: wp("5%"),
+    height: hp("33%"),
+    width: wp("85%"),
+    marginLeft: wp("2.5%"),
+    borderRadius: 10,
+    backgroundColor: "white",
+    fontSize: hp("2%"),
+    fontFamily: "Roboto-Light",
+  },
+  nickname: {
+    paddingLeft: wp("5%"),
+    height: hp("7%"),
+    width: wp("85%"),
+    marginLeft: wp("2.5%"),
+    marginBottom: hp("2%"),
+    borderRadius: 10,
+    backgroundColor: "white",
+    fontSize: hp("2%"),
+    fontFamily: "Roboto-Light",
+  },
+  commentHeadingContainer: {
+    marginLeft: wp("2.5%"),
+    marginBottom: hp("0.2%"),
+  },
+  commentHeading: {
+    fontSize: hp("2.5%"),
+    fontFamily: "Roboto-Light",
+  },
+  innerText: {
+    fontSize: hp("1.5%"),
+    fontFamily: "Roboto-Light",
+    marginBottom: hp("0.2%"),
+  },
+  ratingContainer: {
+    marginLeft: wp("25%"),
+    marginRight: wp("25%"),
+    marginBottom: hp("3%"),
+    flexDirection: "row",
+    justifyContent: "space-between",
+  },
+  reviewTitleText: {
+    marginBottom: hp("2%"),
+    fontSize: hp("4%"),
+    fontFamily: "Roboto-Bold",
+    textAlign: "center",
+  },
+  editButton: {
+    height: hp("5%"),
+    borderBottomLeftRadius: 10,
+    borderBottomRightRadius: 10,
+    textAlign: "center",
+    backgroundColor: "#343239",
+    borderColor: "#ededed",
+  },
+  editButtonText: {
+    marginTop: hp("0.75%"),
+    color: "#ededed",
+    fontSize: hp("2.5%"),
+    fontWeight: "bold",
+    textAlign: "center",
+  },
+});
+
+export const premiumModalStyles = StyleSheet.create({
+  blurView: {
+    width: wp("100%"),
+    height: hp("100%"),
+  },
+  modalWrap: {
+    alignSelf: "center",
+    marginTop: hp("1%"),
+  },
+  modal: {
+    borderWidth: 0,
+    borderColor: "#d6d6d6",
+  },
+  modalInnerWrap: {
+    marginTop: hp("3%"),
+    alignSelf: "center",
+    width: wp("90%"),
+    height: hp("90%"),
+    borderRadius: 5,
+  },
+  modalInnerWrapComment: {
+    backgroundColor: "#ededed",
+    marginTop: hp("3%"),
+    alignSelf: "center",
+    width: wp("90%"),
+    height: hp("89%"),
+    borderRadius: 5,
+  },
+  closeButtonContainer: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+  },
+  closeButton: {
+    marginTop: hp("0.5%"),
+    marginRight: hp("0.5%"),
+  },
+  titleContainer: {
+    backgroundColor: "#1e1c24",
+    borderRadius: 10,
+    width: wp("86%"),
+    minHeight: hp("5.25%"),
+    marginLeft: wp("2.5%"),
+    marginRight: wp("2.5%"),
+    marginBottom: hp("0.25%"),
+    textAlign: "center",
+    flexDirection: "row",
+    flex: 1,
+    justifyContent: "space-between",
+  },
+  title: {
+    fontSize: hp("2.5%"),
+    color: "#ededed",
+    fontFamily: "Roboto-Bold",
+    marginTop: hp("0.65%"),
+    marginLeft: wp("2%"),
+  },
+  priceContainer: {
+    height: hp("3%"),
+    width: wp("20%"),
+    marginTop: hp("0.9%"),
+    marginRight: wp("1.5%"),
+    borderRadius: 5,
+  },
+  price: {
+    marginTop: hp("0.25%"),
+    color: "#ededed",
+    fontSize: hp("1.75%"),
+    fontFamily: "Roboto-Bold",
+    textAlign: "center",
+  },
+  centerdWrap: {
+    alignSelf: "center",
+  },
+  basicUserInfo: {
+    maxHeight: hp("50%"),
+    marginBottom: wp("1%"),
+    marginTop: hp("1%"),
+    marginLeft: wp("2.5%"),
+    marginRight: wp("2.5%"),
+    backgroundColor: "#ffffff",
+    borderRadius: 10,
+  },
+  profileImageBorderHeight: {
+    alignSelf: "center",
+    marginTop: hp("1.5%"),
+    width: hp("16%"),
+    height: hp("16%"),
+    borderWidth: 1,
+    borderColor: "#1e1c24",
+  },
+  profileImageBorderWidth: {
+    alignSelf: "center",
+    marginTop: wp("3%"),
+    width: wp("32%"),
+    height: wp("32%"),
+    borderWidth: 1,
+    borderColor: "#1e1c24",
+  },
+  mainContainer: {
+    alignSelf: "center",
+    marginTop: hp("3%"),
+    width: wp("90%"),
+    height: wp("60%"),
+  },
+  description: {
+    minHeight: hp("37%"),
+    marginBottom: hp("0.5%"),
+    marginTop: hp("1%"),
+    marginLeft: wp("2.5%"),
+    marginRight: wp("2.5%"),
+    paddingLeft: wp("2%"),
+    paddingLeft: wp("1%"),
+    paddingTop: wp("2%"),
+    paddingBottom: wp("2%"),
+    backgroundColor: "#ffffff",
+    borderRadius: 10,
+  },
+  profileImageHeight: {
+    alignSelf: "center",
+    marginTop: hp("0.35%"),
+    width: hp("15%"),
+    height: hp("15%"),
+  },
+  profileImageWidth: {
+    alignSelf: "center",
+    marginTop: wp("0.7%"),
+    width: wp("30%"),
+    height: wp("30%"),
+  },
+  profileName: {
+    alignSelf: "center",
+    marginTop: hp("0.25%"),
+    marginBottom: hp("0.5%"),
+    fontSize: hp("2.75%"),
+    fontWeight: "bold",
+    fontFamily: "Roboto-Light",
+    color: "#000000",
+  },
+  userLocation: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginBottom: hp("0.25%"),
+  },
+  location: {
+    marginLeft: wp("1%"),
+    fontSize: wp("4%"),
+    fontFamily: "Roboto-Light",
+    color: "#000000",
+  },
+  userMail: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginBottom: hp("0.5%"),
+  },
+  userOntact: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginBottom: hp("1.5%"),
+  },
+  userRating: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginBottom: hp("1.5%"),
+  },
+  ratingText: {
+    fontSize: hp("2%"),
+    fontFamily: "Roboto-Light",
+    color: "#000000",
+    marginLeft: wp("1%"),
+    marginRight: wp("3%"),
+  },
+  likeComponent: {
+    color: "#000000",
+  },
+  like: {
+    fontSize: hp("2.75%"),
+    color: "#000000",
+  },
+  dislikeComponent: {
+    color: "#000000",
+  },
+  dislike: {
+    fontSize: hp("2.75%"),
+    color: "#000000",
+  },
+  textInput: {
+    padding: wp("5%"),
+    height: hp("33%"),
+    width: wp("85%"),
+    marginLeft: wp("2.5%"),
+    borderRadius: 10,
+    backgroundColor: "white",
+    fontSize: hp("2%"),
+    fontFamily: "Roboto-Light",
+  },
+  nickname: {
+    paddingLeft: wp("5%"),
+    height: hp("7%"),
+    width: wp("85%"),
+    marginLeft: wp("2.5%"),
+    marginBottom: hp("2%"),
+    borderRadius: 10,
+    backgroundColor: "white",
+    fontSize: hp("2%"),
+    fontFamily: "Roboto-Light",
+  },
+  commentHeadingContainer: {
+    marginLeft: wp("2.5%"),
+    marginBottom: hp("0.2%"),
+  },
+  commentHeading: {
+    fontSize: hp("2.5%"),
+    fontFamily: "Roboto-Light",
+  },
+  innerText: {
+    fontSize: hp("1.5%"),
+    fontFamily: "Roboto-Light",
+    marginBottom: hp("0.2%"),
+  },
+  ratingContainer: {
+    marginLeft: wp("25%"),
+    marginRight: wp("25%"),
+    marginBottom: hp("3%"),
+    flexDirection: "row",
+    justifyContent: "space-between",
+  },
+  reviewTitleText: {
+    marginBottom: hp("2%"),
+    fontSize: hp("4%"),
+    fontFamily: "Roboto-Bold",
+    textAlign: "center",
+  },
+  editButton: {
+    height: hp("5%"),
+    borderBottomLeftRadius: 10,
+    borderBottomRightRadius: 10,
+    textAlign: "center",
+    backgroundColor: "#343239",
+    borderColor: "#ededed",
+  },
+  editButtonText: {
+    marginTop: hp("0.75%"),
+    color: "#ededed",
+    fontSize: hp("2.5%"),
+    fontWeight: "bold",
+    textAlign: "center",
+  },
+});
