+++ react_setup/AwesomeProject/shared/profileStyles.js
@@ -1,0 +1,274 @@
+import { Platform, StyleSheet } from "react-native";
+import { withTheme } from "react-native-elements";
+import {
+  widthPercentageToDP as wp,
+  heightPercentageToDP as hp,
+} from "react-native-responsive-screen";
+
+const customFonts = {
+  "Comfortaa-Regular": require("../assets/fonts/Comfortaa-Regular.ttf"),
+  "Comfortaa-Light": require("../assets/fonts/Comfortaa-Light.ttf"),
+  "Comfortaa-Bold": require("../assets/fonts/Comfortaa-Bold.ttf"),
+  "Comfortaa-Bold": require("./../assets/fonts/Comfortaa-Bold.ttf"),
+  "Roboto-Thin": require("../assets/fonts/Roboto-Thin.ttf"),
+  "Roboto-Bold": require("../assets/fonts/Roboto-Bold.ttf"),
+  "Roboto-Regular": require("../assets/fonts/Roboto-Regular.ttf"),
+  "Roboto-Light": require("../assets/fonts/Roboto-Light.ttf"),
+};
+
+export const profileStyles = StyleSheet.create({
+  backgroundImageContainer: {
+    flex: 1,
+    resizeMode: "cover",
+  },
+  mainContainer: {
+    alignSelf: "center",
+    marginTop: hp("3%"),
+    width: wp("90%"),
+    minHeight: hp("70%"),
+    marginBottom: hp("1%"),
+  },
+  editButton: {
+    height: hp("5%"),
+    borderBottomLeftRadius: 10,
+    borderBottomRightRadius: 10,
+    textAlign: "center",
+    backgroundColor: "#03d5ff",
+    borderColor: "#ededed",
+  },
+  editButtonText: {
+    marginTop: hp("1%"),
+    color: "#ededed",
+    fontSize: wp("4%"),
+    fontWeight: "bold",
+    textAlign: "center",
+  },
+  basicUserInfo: {
+    maxHeight: hp("50%"),
+    marginBottom: hp("1%"),
+    backgroundColor: "#2d2d2d",
+    borderRadius: 10,
+    borderWidth: 1,
+    borderColor: "#ededed",
+    shadowColor: "#000",
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+
+    elevation: 5,
+  },
+  profileImageHeight: {
+    alignSelf: "center",
+    marginTop: hp("1.5%"),
+    width: hp("15%"),
+    height: hp("15%"),
+    borderRadius: 100,
+  },
+  profileImageWidth: {
+    alignSelf: "center",
+    marginTop: wp("3%"),
+    width: wp("30%"),
+    height: wp("30%"),
+    borderRadius: 100,
+  },
+  titleIconWidth: {
+    width: wp("12%"),
+    height: wp("12%"),
+    top: wp("4%"),
+    left: wp("2.5%"),
+  },
+  profileName: {
+    alignSelf: "center",
+    marginTop: hp("0.25%"),
+    marginBottom: hp("0.25%"),
+    fontSize: wp("4.75%"),
+    fontWeight: "bold",
+    color: "#ededed",
+    fontFamily: "Roboto-Bold",
+  },
+  userLocation: {
+    flexDirection: "row",
+    alignSelf: "center",
+    marginBottom: hp("0.25%"),
+  },
+  location: {
+    marginLeft: wp("1%"),
+    fontSize: wp("4%"),
+    fontFamily: "Roboto-Light",
+    color: "#ededed",
+  },
+  userMail: {
+    flexDirection: "row",
+    alignSelf: "center",
+    fontFamily: "Roboto-Light",
+    marginBottom: hp("0.5%"),
+  },
+  userOntact: {
+    flexDirection: "row",
+    alignSelf: "center",
+    fontFamily: "Roboto-Light",
+    marginBottom: hp("0.5%"),
+  },
+  userRating: {
+    flexDirection: "row",
+    alignSelf: "center",
+    justifyContent: "space-between",
+    marginBottom: hp("1.5%"),
+  },
+  ratingText: {
+    fontSize: wp("3.5%"),
+    fontFamily: "Roboto-Light",
+    color: "#ededed",
+    marginLeft: wp("1%"),
+    marginRight: wp("3%"),
+  },
+  likeComponent: {
+    color: "#ededed",
+  },
+  like: {
+    fontSize: wp("4.8%"),
+    color: "#ededed",
+  },
+  dislikeComponent: {
+    color: "#ededed",
+  },
+  dislike: {
+    fontSize: wp("4.8%"),
+    color: "#ededed",
+  },
+  userContact: {
+    flexDirection: "row",
+    alignSelf: "center",
+  },
+  userEmail: {
+    flexDirection: "row",
+    alignSelf: "center",
+  },
+  aboutUser: {
+    marginBottom: hp("1%"),
+    backgroundColor: "#2d2d2d",
+    borderRadius: 10,
+    borderWidth: 1,
+    borderColor: "#ededed",
+    shadowColor: "#000",
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+
+    elevation: 5,
+  },
+  smallContainer: {
+    marginBottom: hp("1%"),
+    backgroundColor: "#2d2d2d",
+    borderRadius: 10,
+    borderWidth: 1,
+    borderColor: "#ededed",
+    shadowColor: "#000",
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+
+    elevation: 5,
+  },
+  smallContainerBottom: {
+    marginBottom: hp("5%"),
+    backgroundColor: "#ededed",
+    borderRadius: 10,
+    borderWidth: 1,
+    borderColor: "#606060",
+  },
+  sectionName: {
+    left: wp("2%"),
+    marginTop: hp("1.5%"),
+    marginBottom: hp("1.5%"),
+    marginLeft: wp("2%"),
+    marginRight: wp("2%"),
+    fontSize: wp("4.25%"),
+    textAlign: "center",
+    fontFamily: "Roboto-Bold",
+    color: "#ededed",
+  },
+  dashContainer: {
+    marginLeft: wp("0.3%"),
+    flexDirection: "row",
+  },
+  line: {
+    textAlignVertical: "center",
+    fontSize: wp("5.25%"),
+    color: "black",
+  },
+  userDetails: {
+    paddingHorizontal: hp("1.25%"),
+    paddingVertical: wp("2%"),
+    marginHorizontal: wp("1.5%"),
+    marginVertical: hp("0.5%"),
+    borderRadius: 10,
+    backgroundColor: "#1e1c24",
+    borderWidth: wp("0.1%"),
+    borderColor: "#9f9f9f",
+  },
+  comment: {
+    paddingHorizontal: hp("1.25%"),
+    paddingVertical: wp("2%"),
+    marginHorizontal: wp("1.5%"),
+    marginVertical: hp("0.5%"),
+    borderRadius: 10,
+    backgroundColor: "white",
+  },
+  details: {
+    fontSize: wp("3.5%"),
+    fontFamily: "Roboto-Light",
+    lineHeight: wp("4%"),
+    color: "#ededed",
+  },
+  commentTitle: {
+    fontSize: wp("3.5%"),
+    fontWeight: "bold",
+    color: "#ededed",
+  },
+  commentButtonText: {
+    fontSize: wp("4.25%"),
+    textAlign: "center",
+    fontWeight: "bold",
+    color: "black",
+  },
+  commentUser: {
+    fontSize: wp("3.5%"),
+    fontWeight: "bold",
+    marginTop: hp("0.5%"),
+    marginLeft: wp("2%"),
+    color: "#ededed",
+  },
+  commentTime: {
+    fontSize: wp("3.5%"),
+    marginTop: hp("0.5%"),
+    marginLeft: wp("1%"),
+    color: "#ededed",
+  },
+  commentText: {
+    fontSize: wp("3.5%"),
+    marginTop: hp("1%"),
+    marginLeft: wp("2%"),
+    color: "#ededed",
+  },
+  arrow: {
+    alignSelf: "center",
+    fontSize: wp("5.25%"),
+    color: "#ededed",
+  },
+  arrowSmall: {
+    alignSelf: "center",
+    marginTop: hp("0.5%"),
+    fontSize: wp("5.25%"),
+    color: "#ededed",
+  },
+});
